// src/app/models/abstract/entryShipment.interface.ts
import { Shipment } from "../shipment.model";
import { SelectItem } from "primeng/primeng";
import { FormControl } from "@angular/forms";
import { Customer } from "..";
import { SelectModel } from "../select.model";

export interface EntryShipment extends Shipment {
    dim?: number;
    priceList?: SelectItem[];
    cloneCustomer: Customer;
    filteredToHubs?: any[];
    toHubLabel?: string;
    toCenterHubId?: number;
    deliveryUsers?: SelectItem[];
    deliveryUserLabel?: string;
    filteredDeliveryUsers?: any[];
    pickupUsers?: SelectItem[];
    pickupUserLabel?: string;
    filteredPickupUsers?: any[];
    customers?: SelectModel[];
    customerLabel?: string;
    filteredCustomers?: any[];
    structures?: SelectItem[];
    structureLabel?: string;
    filteredStructures?: any[];
    paymentTypes?: SelectItem[];
    paymentTypeLabel: string;
    filteredPaymentTypes?: any[];
    cusDepartments?: SelectItem[];
    cusDepartmentLabel?: string;
    filteredCusDepartments?: any[];
    filteredServices?: any[];
    msgService?: string;
    services?: SelectItem[];
    serviceLabel?: string;
    customer?: Customer;
    toHubs?: SelectModel[];
    fromHubs?: SelectModel[];
    isCreateMissingInfo?: boolean;
    isShowPrice?: boolean;
    isShowDVGTs?: boolean;
    toProvinceLabel?: string;
    fromProvinceLabel?: string;
    filteredToProvinces?: any[];
    toDistricts?: SelectModel[];
    fromDistricts?: SelectModel[];
    toDistrictLabel?: string;
    fromDistrictLabel?: string;
    filteredToDistricts?: any[];
    toWards?: SelectModel[];
    fromWards?: SelectModel[];
    filteredToWards?: any[];
    toWardLabel?: string;
    fromWardLabel?: string;
}